LayoutTests/accessibility/img-aria-button-alt-tag.html

<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>
<body id="body">
 
    <img id='image' src="resources/cake.png" role="button" alt="alternate" title="title">
    
    <p id="description"></p>
    <div id="console"></div>
     
    <script>
        description("This test makes sure that a img acting as an aria button still returns its alt tag as the accessible name");
 
        if (window.accessibilityController) {
            var result = document.getElementById("result");

            var body = document.getElementById("body").focus();
            var imgUIElement = accessibilityController.focusedElement.childAtIndex(0).childAtIndex(0);
            shouldBeEqualToString("platformValueForW3CName(imgUIElement)", "alternate");
        }
    </script>
    
<script src="../resources/js-test-post.js"></script>
</body>
</html>
