Safari-611.1.10/LayoutTests/accessibility/img-aria-button-alt-tag-expected.txt


This test makes sure that a img acting as an aria button still returns its alt tag as the accessible name

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS platformValueForW3CName(imgUIElement) is "alternate"
PASS successfullyParsed is true

TEST COMPLETE
